+++ src/main/kotlin/joshxviii/plantz/entity/plant/Explosive.kt
@@ -1,5 +1,0 @@
-package joshxviii.plantz.entity.plant
-
-interface Explosive {
-
-}